[*] Rework of data formatting services, [+] Added input for id in homepage
