+++ sportsee/src/App.js
@@ -1,9 +1,10 @@
 import { Routes, Route } from "react-router-dom";
 import "./css/compiled/compiled.css";
 import ProfilPage from "./render/profilPage";
-import ErrorPage from "./render/components/errorPage";
+import ErrorPage from "./render/errorPage";
 import HorizontalNav from "./render/components/horizontalNav";
 import VerticalNav from "./render/components/verticalNav";
+import HomePage from "./render/homePage";
 
 function App() {
     return (
@@ -11,6 +12,7 @@ function App() {
             <HorizontalNav />
             <VerticalNav />
             <Routes>
+                <Route path="/" element={<HomePage />} />
                 <Route path="/profil/:id" element={<ProfilPage />} />
                 <Route path="*" element={<ErrorPage title={"404"} content={"La page que vous recherchez n'existe pas..."} />} />
             </Routes>
