+++ sportsee/src/services/activityDataService.js
@@ -1,0 +1,17 @@
+class ActivityDataService {
+    static formatActivityData(activityData) {
+        let formattedData = [];
+
+        activityData.data.sessions.map((activity) => {
+            return formattedData.push({
+                ...activity,
+                "tooltipKilogram": `${activity.kilogram}kg`,
+                "tooltipCalories": `${activity.calories}Kcal`
+            });
+        })
+
+        return formattedData;
+    }
+}
+
+export default ActivityDataService;
