+++ sportsee/src/services/nameDataService.js
@@ -1,0 +1,10 @@
+class NameDataService {
+    // Firstname of user data formatting
+    static formatNameData(userData) {     
+        return {
+            "firstName": userData.data.userInfos.firstName,
+        };
+    }
+}
+
+export default NameDataService;
